Cast when printing dispatch_unote_ident_t.
Different platforms may define dispatch_unote_ident_t differently
and subsequently this type may have different bit widths. Therefore,
when printing du_ident, cast explicitly to deal with format mismatch
errors.

The alternative is to specify explicit format string macros, but
this is a little complex and casting is already being used in #584.

One special case for consideration is `_dispatch_timer_unote_disarm`;
this gets the `du_ident` and converts that into a `uint32_t` timer
index. When `dispatch_unote_ident` is a `uint32_t` this is of course
fine, but needs consideration for when it isn't. Here, we just cast
down. This is somewhat reasonable since this is initialized from
`_dispatch_timer_unote_idx` which returns an `unsigned int`. (Of course,
we are not actually bounds checking that index, but that's outside the
scope of this commit).

+++ src/init.c
@@ -1049,7 +1049,7 @@ _dispatch_bug_kevent_vanished(dispatch_unote_t du)
 			"{ %p[%s], ident: %" PRIdPTR " / 0x%" PRIxPTR ", handler: %p }",
 			dux_type(du._du)->dst_kind, dou._dq,
 			dou._dq->dq_label ? dou._dq->dq_label : "<unknown>",
-			du._du->du_ident, du._du->du_ident, func);
+			(intptr_t)du._du->du_ident, (uintptr_t)du._du->du_ident, func);
 }
 
 #endif // RDAR_49023449
@@ -1155,7 +1155,7 @@ _dispatch_logv_init(void *context DISPATCH_UNUSED)
 #else
 			dprintf(dispatch_logfile, "=== log file opened for %s[%u] at "
 					"%ld.%06u ===\n", getprogname() ?: "", getpid(),
-					tv.tv_sec, (int)tv.tv_usec);
+					(long)tv.tv_sec, (int)tv.tv_usec);
 #endif
 		}
 	}
